image_recognition/requirements.txt
flask==2.0.1
Werkzeug==2.0.1
Pillow==9.0.0
tensorflow==2.8.0
numpy==1.22.0
gunicorn==20.1.0